[feat] changed parameter type from int to float and improved the visibility of updated records for better clarity.

+++ dev/sjkim/highscore.h
@@ -1,6 +1,6 @@
 #ifndef MODULE3_H
 #define MODULE3_H
 
-void highscore(char[], int);
+void highscore(char[], float);
 
 #endif
